third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/scroll-state/scroll-state-scroll-direction-change.html

<!DOCTYPE html>
<title>@container: scroll-state(scroll-direction) changed after scroll</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/6400">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  #scroller {
    width: 200px;
    height: 200px;
    container-type: scroll-state;
    overflow-x: scroll;
    overflow-y: scroll;
  }
  #filler {
    height: 600px;
    width: 600px;
  }
  #target {
    --top: no;
    --bottom: no;
    --y: no;
    --left: no;
    --right: no;
    --x: no;
    --none: no;
    @container scroll-state(scroll-direction: top) {
      --top: yes;
    }
    @container scroll-state(scroll-direction: bottom) {
      --bottom: yes;
    }
    @container scroll-state(scroll-direction: y) {
      --y: yes;
    }
    @container scroll-state(scroll-direction: left) {
      --left: yes;
    }
    @container scroll-state(scroll-direction: right) {
      --right: yes;
    }
    @container scroll-state(scroll-direction: x) {
      --x: yes;
    }
    @container scroll-state(scroll-direction: none) {
      --none: yes;
    }
  }
</style>
<div id="scroller">
  <div id="filler">
    <div id="target"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--none"), "yes");

    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "no");

    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "no");
  }, "Check that scroll-state(scroll-direction) state before scrolling");

  promise_test(async t => {
    scroller.scrollTop = 300;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
    assert_equals(scroller.scrollTop, 300);
  }, "scroll-state(scroll-direction) after scrolling bottom");

  promise_test(async t => {
    scroller.scrollTop = 100;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--top"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--bottom"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--y"), "yes");
    assert_equals(scroller.scrollTop, 100);
  }, "scroll-state(scroll-direction) after scrolling top");

  promise_test(async t => {
    scroller.scrollLeft = 300;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "yes");
    assert_equals(scroller.scrollLeft, 300);
  }, "scroll-state(scroll-direction) after scrolling right");

  promise_test(async t => {
    scroller.scrollLeft = 100;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--left"), "yes");
    assert_equals(getComputedStyle(target).getPropertyValue("--right"), "no");
    assert_equals(getComputedStyle(target).getPropertyValue("--x"), "yes");
    assert_equals(scroller.scrollLeft, 100);
  }, "scroll-state(scroll-direction) after scrolling left");

</script>
